Fe/feat/booking#63

See merge request s09-final/S09P31C206!274

+++ frontend/app/src/main/java/com/ssafy/booking/ui/booking/bookingSetting/BookingSetLocation.kt
@@ -146,8 +146,8 @@ fun SearchResult(bookingViewModel: BookingViewModel,locationViewModel: LocationV
                                 bookingViewModel.lgt.value = item.x
                                 bookingViewModel.placeName.value = item.placeName
                                 showSearchResults.value = false
-                                App.prefs.putMeetingLat(item.y.toDouble())
-                                App.prefs.putMeetingLgt(item.x.toDouble())
+                                App.prefs.putMeetingLat(item.y.toFloat())
+                                App.prefs.putMeetingLgt(item.x.toFloat())
                                 App.prefs.putMeetingAddress(item.addressName)
                                 App.prefs.putMeetingLocation(item.placeName)
                             },
